4.21*: bloglist expansion, step9

+++ part-4/blog-list/controllers/blogs.js
@@ -37,9 +37,27 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id)
+  const token = req.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
 
-  res.status(204).end()
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(req.params.id)
+  if (!blog || blog.user == undefined) {
+    return res.status(404).end()
+  }
+  const usersBlog = blog.user.toString() === decodedToken.id
+  if (usersBlog) {
+    await Blog.findByIdAndDelete(req.params.id)
+    res.status(204).end()
+  } else {
+    res
+      .status(403)
+      .json({ error: 'invalid user' })
+      .end()
+  }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
